+++ sdk/resourcemanager/hdinsightcontainers/armhdinsightcontainers/fake/clusters_server.go
@@ -29,7 +29,7 @@ type ClustersServer struct {
 	BeginCreate func(ctx context.Context, resourceGroupName string, clusterPoolName string, clusterName string, hdInsightCluster armhdinsightcontainers.Cluster, options *armhdinsightcontainers.ClustersClientBeginCreateOptions) (resp azfake.PollerResponder[armhdinsightcontainers.ClustersClientCreateResponse], errResp azfake.ErrorResponder)
 
 	// BeginDelete is the fake for method ClustersClient.BeginDelete
-	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted, http.StatusNoContent
+	// HTTP status codes to indicate success: http.StatusAccepted, http.StatusNoContent
 	BeginDelete func(ctx context.Context, resourceGroupName string, clusterPoolName string, clusterName string, options *armhdinsightcontainers.ClustersClientBeginDeleteOptions) (resp azfake.PollerResponder[armhdinsightcontainers.ClustersClientDeleteResponse], errResp azfake.ErrorResponder)
 
 	// Get is the fake for method ClustersClient.Get
@@ -59,6 +59,10 @@ type ClustersServer struct {
 	// BeginUpdate is the fake for method ClustersClient.BeginUpdate
 	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted
 	BeginUpdate func(ctx context.Context, resourceGroupName string, clusterPoolName string, clusterName string, clusterPatchRequest armhdinsightcontainers.ClusterPatch, options *armhdinsightcontainers.ClustersClientBeginUpdateOptions) (resp azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpdateResponse], errResp azfake.ErrorResponder)
+
+	// BeginUpgrade is the fake for method ClustersClient.BeginUpgrade
+	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted
+	BeginUpgrade func(ctx context.Context, resourceGroupName string, clusterPoolName string, clusterName string, clusterUpgradeRequest armhdinsightcontainers.ClusterUpgrade, options *armhdinsightcontainers.ClustersClientBeginUpgradeOptions) (resp azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpgradeResponse], errResp azfake.ErrorResponder)
 }
 
 // NewClustersServerTransport creates a new instance of ClustersServerTransport with the provided implementation.
@@ -74,6 +78,7 @@ func NewClustersServerTransport(srv *ClustersServer) *ClustersServerTransport {
 		newListServiceConfigsPager:    newTracker[azfake.PagerResponder[armhdinsightcontainers.ClustersClientListServiceConfigsResponse]](),
 		beginResize:                   newTracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientResizeResponse]](),
 		beginUpdate:                   newTracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpdateResponse]](),
+		beginUpgrade:                  newTracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpgradeResponse]](),
 	}
 }
 
@@ -88,6 +93,7 @@ type ClustersServerTransport struct {
 	newListServiceConfigsPager    *tracker[azfake.PagerResponder[armhdinsightcontainers.ClustersClientListServiceConfigsResponse]]
 	beginResize                   *tracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientResizeResponse]]
 	beginUpdate                   *tracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpdateResponse]]
+	beginUpgrade                  *tracker[azfake.PollerResponder[armhdinsightcontainers.ClustersClientUpgradeResponse]]
 }
 
 // Do implements the policy.Transporter interface for ClustersServerTransport.
@@ -120,6 +126,8 @@ func (c *ClustersServerTransport) Do(req *http.Request) (*http.Response, error) 
 		resp, err = c.dispatchBeginResize(req)
 	case "ClustersClient.BeginUpdate":
 		resp, err = c.dispatchBeginUpdate(req)
+	case "ClustersClient.BeginUpgrade":
+		resp, err = c.dispatchBeginUpgrade(req)
 	default:
 		err = fmt.Errorf("unhandled API %s", method)
 	}
@@ -220,9 +228,9 @@ func (c *ClustersServerTransport) dispatchBeginDelete(req *http.Request) (*http.
 		return nil, err
 	}
 
-	if !contains([]int{http.StatusOK, http.StatusAccepted, http.StatusNoContent}, resp.StatusCode) {
+	if !contains([]int{http.StatusAccepted, http.StatusNoContent}, resp.StatusCode) {
 		c.beginDelete.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted, http.StatusNoContent", resp.StatusCode)}
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusAccepted, http.StatusNoContent", resp.StatusCode)}
 	}
 	if !server.PollerResponderMore(beginDelete) {
 		c.beginDelete.remove(req)
@@ -535,6 +543,58 @@ func (c *ClustersServerTransport) dispatchBeginUpdate(req *http.Request) (*http.
 	}
 	if !server.PollerResponderMore(beginUpdate) {
 		c.beginUpdate.remove(req)
+	}
+
+	return resp, nil
+}
+
+func (c *ClustersServerTransport) dispatchBeginUpgrade(req *http.Request) (*http.Response, error) {
+	if c.srv.BeginUpgrade == nil {
+		return nil, &nonRetriableError{errors.New("fake for method BeginUpgrade not implemented")}
+	}
+	beginUpgrade := c.beginUpgrade.get(req)
+	if beginUpgrade == nil {
+		const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.HDInsight/clusterpools/(?P<clusterPoolName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/clusters/(?P<clusterName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/upgrade`
+		regex := regexp.MustCompile(regexStr)
+		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
+		if matches == nil || len(matches) < 4 {
+			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
+		}
+		body, err := server.UnmarshalRequestAsJSON[armhdinsightcontainers.ClusterUpgrade](req)
+		if err != nil {
+			return nil, err
+		}
+		resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
+		if err != nil {
+			return nil, err
+		}
+		clusterPoolNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("clusterPoolName")])
+		if err != nil {
+			return nil, err
+		}
+		clusterNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("clusterName")])
+		if err != nil {
+			return nil, err
+		}
+		respr, errRespr := c.srv.BeginUpgrade(req.Context(), resourceGroupNameParam, clusterPoolNameParam, clusterNameParam, body, nil)
+		if respErr := server.GetError(errRespr, req); respErr != nil {
+			return nil, respErr
+		}
+		beginUpgrade = &respr
+		c.beginUpgrade.add(req, beginUpgrade)
+	}
+
+	resp, err := server.PollerResponderNext(beginUpgrade, req)
+	if err != nil {
+		return nil, err
+	}
+
+	if !contains([]int{http.StatusOK, http.StatusAccepted}, resp.StatusCode) {
+		c.beginUpgrade.remove(req)
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted", resp.StatusCode)}
+	}
+	if !server.PollerResponderMore(beginUpgrade) {
+		c.beginUpgrade.remove(req)
 	}
 
 	return resp, nil